load: Add optional measurement to gather Cilium Agent pprofs
This commit adds a new module to the load test to gather pprofs for
cilium agents running in the target cluster through the use of
cilium-bugtool, a CLI tool included in all cilium-agent containers. The
bugtool outputs a tarball to stdout containing CPU, heap and trace
pprofs as well as information gathered via gops. This tarball is saved
by the PodPeriodicCommand measurement and can be inspected for
debugging.

Example usage:

```
cat PodPeriodicCommand*cilium*stdout*.txt | tar -xf - -i && \
  go tool pprof -http localhost:8080 cilium-bugtool*/cmd/pprof-cpu
```

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>